[ref] move skeletons to a new directory in components
