+++ ui-new/components/skeletons/challenge.tsx
